Column Numbers on Bitmap and Cleanup MIPS Code
Using Excel to plot all pixel index values throughout Bitmap Display, the values needed for the column numbers on the Bitmap Display were collected to generate a list of indices to change.  These are listed at the top of the program in a word array.  Referencing this array, a simple loop was create to initialize those specific pixels to the 0x00000000 for white.  Bitmap Display now shows numbers for columns.  Also some minor code cleanup house-keeping so all the code can more easily read.