Added Player Wins to Bitmap and Exception Handling
Player 1 and 2 wins on 2-player mode now print to the Bitmap Display the appropriate win statements.  Also condensed Bitmap Display prints to a specific sub-routine that can be called anywhere in the program.  To add Bitmap Displays, simply add the pixel indices needed to be updated on the Bitmap Display and when calling the sub-routine, $a0, $a1, and $a2 need to be initialized before call with base address of pixel indices array, number of pixels to be updated in the array, and color in 24-bit RGB.

Exception handling for improperly entered characters or return carriage when an integer is expected has also been added to avoid terminating the program due to improper inputs.