Sound Implementation Optimized
For easier implementation of sound, sound loop sub-routine created to handle all music.  After declaring number of notes and values in each array for notes, instrument, duration and volume, the sound can be added to the location in the program needed and the music can be tweaked simply by changing values in the data header.